requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
python-dateutil==2.8.2
mysql-connector-python==8.2.0 